+++ src/Network/Network/Models/PSVirtualNetworkGateway.cs
@@ -72,6 +72,15 @@ namespace Microsoft.Azure.Commands.Network.Models
         [Ps1Xml(Target = ViewControl.Table)]
         public string AdminState { get; set; }
 
+        [Ps1Xml(Target = ViewControl.Table)]
+        public bool AllowRemoteVnetTraffic { get; set; }
+
+        [Ps1Xml(Target = ViewControl.Table)]
+        public bool AllowVirtualWanTraffic { get; set; }
+
+        [Ps1Xml(Label = "AutoScaleConfiguration", Target = ViewControl.Table)]
+        public PSVirtualNetworkGatewayAutoscaleConfiguration AutoScaleConfiguration { get; set; }
+
         [JsonIgnore]
         public string IpConfigurationsText
         {
@@ -124,6 +133,12 @@ namespace Microsoft.Azure.Commands.Network.Models
         public string CustomRoutesText
         {
             get { return JsonConvert.SerializeObject(CustomRoutes, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
+        }
+
+        [JsonIgnore]
+        public string AutoScaleConfigurationText
+        {
+            get { return JsonConvert.SerializeObject(AutoScaleConfiguration, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
         }
     }
 }