+++ src/ServiceBus/ServiceBus.Autorest/test/Set-AzServiceBusGeoDRConfigurationBreakPair.Tests.ps1
@@ -17,12 +17,10 @@ if(($null -eq $TestName) -or ($TestName -contains 'Set-AzServiceBusGeoDRConfigur
 Describe 'Set-AzServiceBusGeoDRConfigurationBreakPair' {
     It 'Break' {
         Set-AzServiceBusGeoDRConfigurationBreakPair -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace -Name $env.alias
-        
+
         while($drConfig.ProvisioningState -ne "Succeeded"){
             $drConfig = Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace
-            if ($TestMode -ne 'playback') {
-                Start-Sleep 10
-            }
+            Start-TestSleep 10
         }
 
         $drConfig.Name | Should -Be $env.alias
@@ -31,12 +29,10 @@ Describe 'Set-AzServiceBusGeoDRConfigurationBreakPair' {
         $drConfig.Role | Should -Be "PrimaryNotReplicating"
 
         $drConfig = New-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace -PartnerNamespace $env.secondaryNamespaceResourceId
-        
+
         while($drConfig.ProvisioningState -ne "Succeeded"){
             $drConfig = Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace
-            if ($TestMode -ne 'playback') {
-                Start-Sleep 10
-            }
+            Start-TestSleep 10
         }
     }
 
@@ -44,12 +40,10 @@ Describe 'Set-AzServiceBusGeoDRConfigurationBreakPair' {
         $drConfig = Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace
 
         Set-AzServiceBusGeoDRConfigurationBreakPair -InputObject $drConfig
-        
+
         do{
             $drConfig = Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace
-            if ($TestMode -ne 'playback') {
-                Start-Sleep 10
-            }
+            Start-TestSleep 10
         } while($drConfig.ProvisioningState -ne "Succeeded")
 
         $drConfig.Name | Should -Be $env.alias
@@ -58,12 +52,10 @@ Describe 'Set-AzServiceBusGeoDRConfigurationBreakPair' {
         $drConfig.Role | Should -Be "PrimaryNotReplicating"
 
         $drConfig = New-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace -PartnerNamespace $env.secondaryNamespaceResourceId
-        
+
         while($drConfig.ProvisioningState -ne "Succeeded"){
             $drConfig = Get-AzServiceBusGeoDRConfiguration -Name $env.alias -ResourceGroupName $env.resourceGroup -NamespaceName $env.primaryNamespace
-            if ($TestMode -ne 'playback') {
-                Start-Sleep 10
-            }
+            Start-TestSleep 10
         }
     }
 }