+++ src/Storage/Storage/Common/StorageExtensions.cs
@@ -32,14 +32,14 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Common
         private const int CopySASLifeTimeInMinutesOauth = 7 * 24 * 60 - 2 * 60;
 
         internal static Uri GenerateUriWithCredentials(
-            this CloudFile file)
+            this CloudFile file, bool disableTrailingDot = false)
         {
             if (null == file)
             {
                 throw new ArgumentNullException("file");
             }
 
-            string sasToken = GetFileSASToken(file);
+            string sasToken = GetFileSASToken(file, disableTrailingDot);
             
             if (string.IsNullOrEmpty(sasToken))
             {
@@ -70,7 +70,7 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Common
             }
         }
 
-        private static string GetFileSASToken(CloudFile file)
+        private static string GetFileSASToken(CloudFile file, bool disableTrailingDot = false)
         {
             if (null == file.ServiceClient.Credentials
                 || file.ServiceClient.Credentials.IsAnonymous
@@ -91,6 +91,12 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Common
                 SharedAccessExpiryTime = DateTime.Now.Add(sasLifeTime),
                 Permissions = SharedAccessFilePermissions.Read,
             };
+
+            // When file path contains trailing dot and disableTrailingDot, need generate sas from path without .
+            if (disableTrailingDot && Util.PathContainsTrailingDot(file.Uri.AbsolutePath))
+            {
+                file = new CloudFile(Util.RemoveFileUriTrailingDot(file.Uri), file.ServiceClient.Credentials);
+            }
 
             return file.GetSharedAccessSignature(policy);
         }