+++ src/Automation/Automation/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 ## Upcoming Release
 
+
+## Version 1.9.2
+* Updated Module operation cmdlets to support Powershell 7.2
+
 ## Version 1.9.1
 * Fixed bug: Runbooks Name Pattern failures.